Pushing changes for 1.4.6.
Added all the currently missing objects that I can manually add, producer machine products are next, added all missing object i18n data, updated crafting recipes and added cooking recipes for the bat wing sandwich, added the portrait for tuxford and fixed all non-npc shops to show a portrait instead of only dialogue, fixed the machine shed four heart event with Hephaestus where it didn't warp him at all and left an error in the SMAPI console, attempting to add MigrateId strings, they don't work as of yet. And removed Walk to the Desert related patches, as Garden Village doesn't require it to be installed but Walk to the Desert requires Garden Village, so it should do all of that on its own.